Fix clock not increasing
The clock did not increase because the insert-after, used as the atom
modifier function tried to swap! app-state :clock when it was called
inside a transaction. This causes the clock to go back to it's previous
state.

Now, the swap-doc! function takes care of increasing the clock for each
document modification.