Fix lines inserted at the end of another line
The retain implementation were wrongly jumping lines. Make it not jump
lines, but instead go to the end of line if it should be entirely
retained.

The insert implementation had to be changed to know if it is in the end
of the line, inserting an empty one after it.